+++ apps/api/src/core/services/error/error.service.ts
@@ -6,19 +6,23 @@ import { SentryEventService } from "@src/core/services/sentry-event/sentry-event
 
 @singleton()
 export class ErrorService {
-  private readonly logger = new LoggerService();
+  private readonly logger = LoggerService.forContext(ErrorService.name);
 
   constructor(
     @InjectSentry() private readonly sentry: Sentry,
     private readonly sentryEventService: SentryEventService
   ) {}
 
-  async execWithErrorHandler<T>(extraLog: Record<string, unknown>, cb: () => Promise<T>): Promise<T> {
+  async execWithErrorHandler<T>(extraLog: Record<string, unknown>, cb: () => Promise<T>, onError?: (error: unknown) => void): Promise<T> {
     try {
       return await cb();
     } catch (error) {
       const sentryEventId = this.sentry.captureEvent(this.sentryEventService.toEvent(error));
       this.logger.error({ error: error.stack, sentryEventId, ...extraLog });
+
+      if (typeof onError === "function") {
+        onError(error);
+      }
     }
   }
 }