+++ apps/api/src/core/services/postgres-logger/postgres-logger.service.ts
@@ -17,7 +17,7 @@ export class PostgresLoggerService implements LogWriter {
 
   constructor(options?: PostgresLoggerServiceOptions) {
     const orm = options?.orm || "drizzle";
-    this.logger = new LoggerService({ context: "POSTGRES", orm, database: options?.database });
+    this.logger = new LoggerService({ base: { context: "POSTGRES", orm, database: options?.database } });
     this.isDrizzle = orm === "drizzle";
     this.useFormat = options?.useFormat || false;
   }