+++ apps/api/src/deployment/services/draining-deployment/draining-deployment.service.ts
@@ -1,9 +1,9 @@
 import { singleton } from "tsyringe";
 
+import { BlockHttpService } from "@src/chain/services/block-http/block-http.service";
 import { DeploymentConfig, InjectDeploymentConfig } from "@src/deployment/config/config.provider";
 import { DrainingDeploymentOutput, LeaseRepository } from "@src/deployment/repositories/lease/lease.repository";
 import { averageBlockCountInAnHour } from "@src/utils/constants";
-import { BlockHttpService } from "../block-http/block-http.service";
 
 @singleton()
 export class DrainingDeploymentService {
@@ -16,8 +16,9 @@ export class DrainingDeploymentService {
   async findDeployments(owner: string, denom: string): Promise<DrainingDeploymentOutput[]> {
     const currentHeight = await this.blockHttpService.getCurrentHeight();
     const closureHeight = Math.floor(currentHeight + averageBlockCountInAnHour * this.config.AUTO_TOP_UP_JOB_INTERVAL_IN_H);
+    const denomAliased = denom === "uakt" ? denom : "uusdc";
 
-    return await this.leaseRepository.findDrainingLeases({ owner, closureHeight, denom });
+    return await this.leaseRepository.findDrainingLeases({ owner, closureHeight, denom: denomAliased });
   }
 
   async calculateTopUpAmount(deployment: Pick<DrainingDeploymentOutput, "blockRate">): Promise<number> {