+++ apps/api/src/deployment/services/top-up-custodial-deployments/top-up-custodial-deployments.service.spec.ts
@@ -1,11 +1,19 @@
+import "@test/mocks/logger-service.mock";
+
 import { AllowanceHttpService, BalanceHttpService, Denom } from "@akashnetwork/http-sdk";
 import { faker } from "@faker-js/faker";
+import { MsgExec } from "cosmjs-types/cosmos/authz/v1beta1/tx";
+import { describe } from "node:test";
+import { container } from "tsyringe";
 
-import { MasterSigningClientService, MasterWalletService } from "@src/billing/services";
+import { BILLING_CONFIG } from "@src/billing/providers";
+import { TYPE_REGISTRY } from "@src/billing/providers/type-registry.provider";
+import { UAKT_TOP_UP_MASTER_WALLET } from "@src/billing/providers/wallet.provider";
+import { MasterSigningClientService, MasterWalletService, RpcMessageService } from "@src/billing/services";
+import { BlockHttpService } from "@src/chain/services/block-http/block-http.service";
 import { ErrorService } from "@src/core/services/error/error.service";
 import { config } from "@src/deployment/config";
 import { LeaseRepository } from "@src/deployment/repositories/lease/lease.repository";
-import { BlockHttpService } from "@src/deployment/services/block-http/block-http.service";
 import { DrainingDeploymentService } from "@src/deployment/services/draining-deployment/draining-deployment.service";
 import { TopUpToolsService } from "@src/deployment/services/top-up-tools/top-up-tools.service";
 import { TopUpCustodialDeploymentsService } from "./top-up-custodial-deployments.service";
@@ -28,7 +36,7 @@ describe(TopUpCustodialDeploymentsService.name, () => {
   };
   const mockMasterSigningClientService = () => {
     return stub<MasterSigningClientService>({
-      execTx: jest.fn()
+      executeTx: jest.fn()
     });
   };
 
@@ -56,6 +64,8 @@ describe(TopUpCustodialDeploymentsService.name, () => {
     allowanceHttpService,
     balanceHttpService,
     drainingDeploymentService,
+    new RpcMessageService(),
+    blockHttpService,
     errorService
   );
 
@@ -65,10 +75,9 @@ describe(TopUpCustodialDeploymentsService.name, () => {
     grantee: string;
     expectedDeploymentsTopUpCount?: 0 | 1 | 2;
     hasDeployments?: boolean;
-    client: MasterSigningClientService;
   };
 
-  const seedFor = ({ denom, balance = "100000000", grantee, expectedDeploymentsTopUpCount = 2, hasDeployments = true, client }: SeedParams) => {
+  const seedFor = ({ denom, balance = "100000000", grantee, expectedDeploymentsTopUpCount = 2, hasDeployments = true }: SeedParams) => {
     const owner = AkashAddressSeeder.create();
 
     return {
@@ -94,42 +103,36 @@ describe(TopUpCustodialDeploymentsService.name, () => {
               isExpectedToTopUp: expectedDeploymentsTopUpCount > 1
             }
           ]
-        : [],
-      client: client
+        : []
     };
   };
 
   const data = [
     seedFor({
       denom: "uakt",
-      grantee: UAKT_TOP_UP_MASTER_WALLET_ADDRESS,
-      client: uaktMasterSigningClientService
+      grantee: UAKT_TOP_UP_MASTER_WALLET_ADDRESS
     }),
     seedFor({
       denom: "ibc/170C677610AC31DF0904FFE09CD3B5C657492170E7E52372E48756B71E56F2F1",
-      grantee: USDT_TOP_UP_MASTER_WALLET_ADDRESS,
-      client: usdtMasterSigningClientService
+      grantee: USDT_TOP_UP_MASTER_WALLET_ADDRESS
     }),
     seedFor({
       denom: "uakt",
       balance: "5500000",
       grantee: UAKT_TOP_UP_MASTER_WALLET_ADDRESS,
-      expectedDeploymentsTopUpCount: 1,
-      client: uaktMasterSigningClientService
+      expectedDeploymentsTopUpCount: 1
     }),
     seedFor({
       denom: "uakt",
       balance: "5500000",
       grantee: UAKT_TOP_UP_MASTER_WALLET_ADDRESS,
-      hasDeployments: false,
-      client: uaktMasterSigningClientService
+      hasDeployments: false
     }),
     seedFor({
       denom: "uakt",
       balance: "0",
       grantee: UAKT_TOP_UP_MASTER_WALLET_ADDRESS,
-      expectedDeploymentsTopUpCount: 0,
-      client: uaktMasterSigningClientService
+      expectedDeploymentsTopUpCount: 0
     })
   ];
 
@@ -154,20 +157,69 @@ describe(TopUpCustodialDeploymentsService.name, () => {
         ?.drainingDeployments?.map(({ deployment }) => deployment) || []
     );
   });
-  jest.spyOn(drainingDeploymentService, "calculateTopUpAmount").mockImplementation(async () => faker.number.int({ min: 2000000, max: 4000000 }));
-  jest.spyOn(topUpDeploymentsService, "topUpDeployment");
+  jest.spyOn(drainingDeploymentService, "calculateTopUpAmount").mockImplementation(async () => faker.number.int({ min: 3500000, max: 4000000 }));
 
   it("should top up draining deployment given owners have sufficient grants and balances", async () => {
-    await topUpDeploymentsService.topUpDeployments();
+    await topUpDeploymentsService.topUpDeployments({ dryRun: false });
 
-    expect(topUpDeploymentsService.topUpDeployment).toHaveBeenCalledTimes(5);
+    expect(uaktMasterSigningClientService.executeTx).toHaveBeenCalledTimes(3);
+    expect(usdtMasterSigningClientService.executeTx).toHaveBeenCalledTimes(2);
 
-    data.forEach(({ drainingDeployments, client }) => {
+    data.forEach(({ drainingDeployments, grant }) => {
       drainingDeployments.forEach(({ isExpectedToTopUp, deployment }) => {
         if (isExpectedToTopUp) {
-          expect(topUpDeploymentsService.topUpDeployment).toHaveBeenCalledWith(expect.any(Number), deployment, client);
+          const client = deployment.denom === "uakt" ? uaktMasterSigningClientService : usdtMasterSigningClientService;
+          expect(client.executeTx).toHaveBeenCalledWith(
+            [
+              {
+                typeUrl: MsgExec.typeUrl,
+                value: {
+                  grantee: grant.grantee,
+                  msgs: [
+                    {
+                      typeUrl: "/akash.deployment.v1beta3.MsgDepositDeployment",
+                      value: expect.any(Buffer)
+                    }
+                  ]
+                }
+              }
+            ],
+            { fee: { granter: grant.granter } }
+          );
         }
       });
+    });
+  });
+
+  xdescribe("actual top up deployment tx on demand", () => {
+    jest.setTimeout(30000);
+
+    it("should top up or not depending on parameters below", async () => {
+      const denom = "uakt";
+      const owner = "<REPLACE_WITH_OWNER_ADDRESS>";
+      const dseq = "<REPLACE_WITH_DEPLOYMENT_DSEQ_NUMBER>" as unknown as number;
+      const wallet = container.resolve<MasterWalletService>(UAKT_TOP_UP_MASTER_WALLET);
+      const signer = new MasterSigningClientService(container.resolve(BILLING_CONFIG), wallet, container.resolve(TYPE_REGISTRY));
+      const grantee = await wallet.getFirstAddress();
+
+      try {
+        const res = await signer.executeTx(
+          [
+            container.resolve(RpcMessageService).getExecDepositDeploymentMsg({
+              dseq,
+              amount: 1000000,
+              denom,
+              owner: owner,
+              grantee
+            })
+          ],
+          { fee: { granter: owner } }
+        );
+
+        console.log("DEBUG res", res);
+      } catch (e) {
+        console.log("DEBUG e", e);
+      }
     });
   });
 });