+++ apps/api/src/deployment/services/top-up-managed-deployments/top-up-managed-deployments.service.spec.ts
@@ -1,47 +1,48 @@
-import { AllowanceHttpService } from "@akashnetwork/http-sdk";
+import "@test/mocks/logger-service.mock";
+
 import { faker } from "@faker-js/faker";
 
-import { MasterSigningClientService, MasterWalletService } from "@src/billing/services";
+import { BillingConfig } from "@src/billing/providers";
+import { UserWalletRepository } from "@src/billing/repositories";
+import { MasterWalletService, RpcMessageService } from "@src/billing/services";
+import { BalancesService } from "@src/billing/services/balances/balances.service";
+import { SimpleSigningStargateClient, TxSignerService } from "@src/billing/services/tx-signer/tx-signer.service";
+import { BlockHttpService } from "@src/chain/services/block-http/block-http.service";
 import { ErrorService } from "@src/core/services/error/error.service";
 import { config } from "@src/deployment/config";
 import { LeaseRepository } from "@src/deployment/repositories/lease/lease.repository";
-import { BlockHttpService } from "@src/deployment/services/block-http/block-http.service";
 import { DrainingDeploymentService } from "@src/deployment/services/draining-deployment/draining-deployment.service";
 import { TopUpManagedDeploymentsService } from "./top-up-managed-deployments.service";
 
 import { AkashAddressSeeder } from "@test/seeders/akash-address.seeder";
-import { DeploymentGrantSeeder } from "@test/seeders/deployment-grant.seeder";
 import { DrainingDeploymentSeeder } from "@test/seeders/draining-deployment.seeder";
+import { UserWalletSeeder } from "@test/seeders/user-wallet.seeder";
 import { stub } from "@test/services/stub";
 
 describe(TopUpManagedDeploymentsService.name, () => {
   const CURRENT_BLOCK_HEIGHT = 7481457;
   const MANAGED_MASTER_WALLET_ADDRESS = AkashAddressSeeder.create();
-  const mockManagedWalletService = (address: string) => {
-    return stub<MasterWalletService>({
-      getFirstAddress: async () => address
-    });
-  };
-  const mockMasterSigningClientService = () => {
-    return stub<MasterSigningClientService>({
-      execTx: jest.fn()
-    });
-  };
-
-  const allowanceHttpService = new AllowanceHttpService();
-  const blockHttpService = stub<BlockHttpService>({ getCurrentHeight: jest.fn() });
-  const managedMasterWalletService = mockManagedWalletService(MANAGED_MASTER_WALLET_ADDRESS);
-  const managedMasterSigningClientService = mockMasterSigningClientService();
-
-  jest.spyOn(blockHttpService, "getCurrentHeight").mockResolvedValue(CURRENT_BLOCK_HEIGHT);
+  const balancesService = stub<BalancesService>({ retrieveAndCalcDeploymentLimit: jest.fn() });
+  const userWalletRepository = stub<UserWalletRepository>({ paginate: jest.fn() });
+  const blockHttpService = stub<BlockHttpService>({ getCurrentHeight: () => CURRENT_BLOCK_HEIGHT });
+  const client = stub<SimpleSigningStargateClient>({ signAndBroadcast: jest.fn() });
+  const txSignerService = stub<TxSignerService>({ getClientForAddressIndex: jest.fn(() => client) });
+  const billingConfig = stub<BillingConfig>({ DEPLOYMENT_GRANT_DENOM: "ibc/170C677610AC31DF0904FFE09CD3B5C657492170E7E52372E48756B71E56F2F1" });
+  const managedMasterWalletService = stub<MasterWalletService>({
+    getFirstAddress: async () => MANAGED_MASTER_WALLET_ADDRESS
+  });
 
   const drainingDeploymentService = new DrainingDeploymentService(blockHttpService, stub<LeaseRepository>(), config);
   const errorService = stub<ErrorService>({ execWithErrorHandler: (params: any, cb: () => any) => cb() });
   const topUpDeploymentsService = new TopUpManagedDeploymentsService(
-    allowanceHttpService,
-    managedMasterWalletService,
-    managedMasterSigningClientService,
+    userWalletRepository,
+    txSignerService,
+    billingConfig,
     drainingDeploymentService,
+    managedMasterWalletService,
+    balancesService,
+    new RpcMessageService(),
+    blockHttpService,
     errorService
   );
 
@@ -52,14 +53,11 @@ describe(TopUpManagedDeploymentsService.name, () => {
   };
 
   const seedFor = ({ balance = "100000000", expectedDeploymentsTopUpCount = 2, hasDeployments = true }: SeedParams) => {
-    const owner = AkashAddressSeeder.create();
+    const wallet = UserWalletSeeder.create();
 
     return {
-      grant: DeploymentGrantSeeder.create({
-        granter: MANAGED_MASTER_WALLET_ADDRESS,
-        grantee: owner,
-        authorization: { spend_limit: { denom: "ibc/170C677610AC31DF0904FFE09CD3B5C657492170E7E52372E48756B71E56F2F1", amount: balance } }
-      }),
+      wallet,
+      balance,
       drainingDeployments: hasDeployments
         ? [
             {
@@ -90,32 +88,44 @@ describe(TopUpManagedDeploymentsService.name, () => {
     seedFor({ balance: "0", expectedDeploymentsTopUpCount: 0 })
   ];
 
-  jest.spyOn(allowanceHttpService, "paginateDeploymentGrants").mockImplementation(async (params, cb) => {
-    return await cb(data.map(({ grant }) => grant));
+  userWalletRepository.paginate.mockImplementation((params, cb) => cb(data.map(({ wallet }) => wallet)));
+
+  jest.spyOn(drainingDeploymentService, "findDeployments").mockImplementation(async owner => {
+    return data.find(({ wallet }) => wallet.address == owner)?.drainingDeployments?.map(({ deployment }) => deployment) || [];
   });
-  jest.spyOn(drainingDeploymentService, "findDeployments").mockImplementation(async (owner, denom) => {
-    return (
-      data
-        .find(({ grant }) => grant.grantee == owner && grant.authorization.spend_limit.denom === denom)
-        ?.drainingDeployments?.map(({ deployment }) => deployment) || []
-    );
+  jest.spyOn(drainingDeploymentService, "calculateTopUpAmount").mockImplementation(async () => faker.number.int({ min: 3500000, max: 4000000 }));
+  balancesService.retrieveAndCalcDeploymentLimit.mockImplementation(async wallet => {
+    return parseInt(data.find(({ wallet: w }) => w.address == wallet.address)?.balance);
   });
-  jest.spyOn(drainingDeploymentService, "calculateTopUpAmount").mockImplementation(async () => faker.number.int({ min: 2000000, max: 4000000 }));
-  jest.spyOn(topUpDeploymentsService, "topUpDeployment");
 
   it("should top up draining deployment given owners have sufficient balances", async () => {
-    await topUpDeploymentsService.topUpDeployments();
+    await topUpDeploymentsService.topUpDeployments({ dryRun: false });
 
     let count = 0;
 
-    data.forEach(({ drainingDeployments }) => {
+    data.forEach(({ wallet, drainingDeployments }) => {
       drainingDeployments.forEach(({ isExpectedToTopUp, deployment }) => {
         if (isExpectedToTopUp) {
-          expect(topUpDeploymentsService.topUpDeployment).toHaveBeenCalledWith(expect.any(Number), deployment);
+          expect(client.signAndBroadcast).toHaveBeenCalledWith([
+            {
+              typeUrl: "/akash.deployment.v1beta3.MsgDepositDeployment",
+              value: {
+                id: {
+                  owner: wallet.address,
+                  dseq: { high: 0, low: deployment.dseq, unsigned: true }
+                },
+                amount: {
+                  denom: deployment.denom,
+                  amount: expect.any(String)
+                },
+                depositor: MANAGED_MASTER_WALLET_ADDRESS
+              }
+            }
+          ]);
           count++;
         }
       });
     });
-    expect(topUpDeploymentsService.topUpDeployment).toHaveBeenCalledTimes(count);
+    expect(client.signAndBroadcast).toHaveBeenCalledTimes(count);
   });
 });