+++ apps/api/test/mocks/logger-service.mock.ts
@@ -1,0 +1,17 @@
+jest.mock("@akashnetwork/logging", () => {
+  class LoggerService {
+    static forContext() {
+      return new LoggerService();
+    }
+
+    error() {}
+    debug() {}
+    info() {}
+  }
+
+  jest.spyOn(LoggerService.prototype, "error");
+  jest.spyOn(LoggerService.prototype, "debug");
+  jest.spyOn(LoggerService.prototype, "info");
+
+  return { LoggerService };
+});